refactored Event into own mod
