+++ src/mqtt/message.rs
@@ -1,9 +1,7 @@
 use tracing::warn;
 
-use crate::{
-    configuration::Configuration,
-    docker::{ContainerEvent, Event, EventType},
-};
+use crate::configuration::Configuration;
+use crate::events::{ContainerEvent, Event, EventType};
 
 use super::{availability, discovery, payload, topic};
