+++ src/mqtt/mod.rs
@@ -4,7 +4,8 @@ use tokio::{
 };
 use tracing::{error, instrument};
 
-use crate::{configuration::Configuration, docker::Event};
+use crate::configuration::Configuration;
+use crate::events::Event;
 
 use self::client::MqttClient;
